Refactor: introduce DTOs for Person/Athlete API separation
Refactored API to use dedicated request/response DTOs for Person and Athlete entities, replacing direct use of domain models. Updated controllers, services, and AutoMapper profiles accordingly. Repository RemoveAsync now returns bool. Improved DI setup with extension methods. Added NotFoundException and enhanced null checks and response codes. Improved code maintainability and clarity.

+++ LipometryAppAPI/Services/AthleteService.cs
@@ -18,29 +18,34 @@ namespace LipometryAppAPI.Services
             _mapper = mapper;
         }
 
-        public async Task<Athlete> CreateAsync(AthleteCreate createAthlete, CancellationToken token = default)
+        public async Task<Athlete> CreateAsync(AthleteCreateRequest createAthlete, CancellationToken token = default)
         {
             await _athleteRepository.CreateAsync(_mapper.Map<Athlete>(createAthlete), token);
             await _unitOfWork.SaveChangesAsync(token);
             return _mapper.Map<Athlete>(createAthlete);
         }
 
-        public async Task<Athlete> UpdateAsync(Guid id, AthleteUpdate updateAthlete, CancellationToken token = default)
+        public async Task<Athlete> UpdateAsync(Guid id, AthleteUpdateRequest updateAthlete, CancellationToken token = default)
         {
-            var existingAthlete = await _athleteRepository.GetByIdAsync(id, token) 
-                ?? throw new Exception("Athlete not found");
+            var existingAthlete = await _athleteRepository.GetByIdAsync(id, token);
+
+            if (existingAthlete is null)
+                return null!;
+
             _mapper.Map(updateAthlete, existingAthlete);
             _athleteRepository.Update(existingAthlete, token);
             await _unitOfWork.SaveChangesAsync(token);
             return existingAthlete;
         }
-        public async Task RemoveAsync(Guid id, CancellationToken token = default)
+        public async Task<bool> RemoveAsync(Guid id, CancellationToken token = default)
         {
-            var existingPerson = await _athleteRepository.GetByIdAsync(id, token)
-                ?? throw new Exception("Athlete not found");
+            var existingPerson = await _athleteRepository.GetByIdAsync(id, token);
 
+            if (existingPerson is null)
+                return false;
             await _athleteRepository.RemoveAsync(id, token);
             await _unitOfWork.SaveChangesAsync(token);
+            return true;
         }
     }
 }
