+++ LipometryAppAPI/Services/PersonService.cs
@@ -1,5 +1,6 @@
 ﻿using AutoMapper;
 using LipometryAppAPI.Contracts.Requests;
+using LipometryAppAPI.Exceptions;
 using LipometryAppAPI.Models;
 using LipometryAppAPI.Repositories;
 
@@ -19,32 +20,34 @@ namespace LipometryAppAPI.Services
             _mapper = mapper;
         }
 
-        public async Task<Person> CreateAsync(PersonCreate createPerson, CancellationToken token = default)
+        public async Task<Person> CreateAsync(PersonCreateRequest createPerson, CancellationToken token = default)
         {
             await _personRepository.CreateAsync(_mapper.Map<Person>(createPerson), token);
             await _unitOfWork.SaveChangesAsync(token);
             return _mapper.Map<Person>(createPerson);
         }
 
-        public async Task<Person> UpdateAsync(Guid id, PersonUpdate updatePerson, CancellationToken token = default)
+        public async Task<Person> UpdateAsync(Guid id, PersonUpdateRequest updatePerson, CancellationToken token = default)
         {
-            var existingPerson = await _personRepository.GetByIdAsync(id, token) 
-                ?? throw new Exception("Person not found");
+            var existingPerson = await _personRepository.GetByIdAsync(id, token);
+
+            if (existingPerson is null)
+                return null!;
 
             _mapper.Map(updatePerson, existingPerson);
             _personRepository.Update(existingPerson, token);
-
             await _unitOfWork.SaveChangesAsync(token);
-
             return existingPerson;
         }
-        public async Task RemoveAsync(Guid id, CancellationToken token = default)
+        public async Task<bool> RemoveAsync(Guid id, CancellationToken token = default)
         {
-            var existingPerson = await _personRepository.GetByIdAsync(id, token)
-                ?? throw new Exception("Person not found");
+            var existingPerson = await _personRepository.GetByIdAsync(id, token);
 
+            if (existingPerson is null)
+                return false;
             await _personRepository.RemoveAsync(id, token);
             await _unitOfWork.SaveChangesAsync(token);
+            return true;
         }
     }
 }
